Create Pagination component
